Refactor OrderItem mapping: update OrderItemMapper to include image mapping and adjust OrderMapper to use OrderItemMapper for response conversion.
